Moved Listeners and Events into their own packages.
Create Player event and listener for use with the character editor.
