+++ src/main/java/net/rpgtoolkit/common/assets/listeners/BoardChangeListener.java
@@ -4,8 +4,9 @@
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
  * the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package net.rpgtoolkit.common.assets;
+package net.rpgtoolkit.common.assets.listeners;
 
+import net.rpgtoolkit.common.assets.events.BoardChangedEvent;
 import java.util.EventListener;
 
 /**
